refactor: implement screens pattern for habit and coach features, and consolidate date-time utils
