+++ .gitignore
@@ -39,3 +39,4 @@ yarn-error.*
 app-example
 /ios
 /android
+todoTask/
