+++ app/(auth)/(tabs)/index.tsx
@@ -28,6 +28,7 @@ import {
 
 import { ScreenView } from "@/components/ui/Themed";
 import ThemeContext from "@/contexts/ThemeContext";
+import { ROUTES } from "@/constants/routes";
 import { useAuth } from "@/contexts/AuthContext";
 import {
   getHabitList,
@@ -134,7 +135,7 @@ export default function TabOneScreen() {
   //   loadHabits(isoDate);
   // }, [isoDate, loadHabits]);
 
-  const onCreateClick = () => router.push("/(auth)/habit/createHabit");
+  const onCreateClick = () => router.push(ROUTES.AUTH.CREATE_HABIT);
 
   const handleHabitDoneClick = async (id: string, count: any) => {
     const currentIsoDate = format(startOfDay(selectedDate), "yyyy-MM-dd");
@@ -217,7 +218,7 @@ export default function TabOneScreen() {
                   <TopBadge
                     iconName="star"
                     variant="pill"
-                    onPress={() => router.push("/(auth)/coach")}
+                    onPress={() => router.push(ROUTES.AUTH.COACH)}
                   />
                 </View>
               )}
