+++ app/(auth)/(tabs)/tools.tsx
@@ -10,6 +10,7 @@ import React, { useContext, useEffect, useState } from "react";
 import { router, useNavigation } from "expo-router";
 // application level import
 import ThemeContext from "@/contexts/ThemeContext";
+import { ROUTES } from "@/constants/routes";
 
 import {
   buttons as NavigationButton,
