--- todoTask/file_str.txt
@@ -1,38 +1,0 @@
-
-  1. Standardize Naming & Pluralization
-  The project currently has a mix of singular/plural folder names and inconsistent casing. Standardizing these will prevent confusion and potential cross-platform issues.
-   * Pluralize root folders: Rename constant/ to constants/ and context/ to contexts/ to match features/, services/, and hooks/.
-   * Standardize Kebab-case: Ensure all directory names use kebab-case (e.g., self-care, check-in, new-user) instead of camelCase or PascalCase.
-   * Route naming: Rename app/(auth)/(tabs)/setting.tsx to settings.tsx to match the settings feature and pluralization standards. Use kebab-case for all route files (e.g., self-care.tsx).
-
-  2. Consolidate Theme & Configuration
-  Theme-related files are currently fragmented across multiple directories.
-   * Merge Theme Constants: Move constant/theme/Colors.ts into the root theme/ directory (e.g., theme/colors.ts).
-   * Unify Types: Consolidate types/themeTypes.ts and theme/types.ts into a single source of truth in theme/types.ts.
-   * Global vs Local constants: Move any feature-specific constants (like builtinAvatars.ts) from the root constant/ folder into features/auth/constants/ or features/user/constants/.
-
-  3. Mature the Feature-Based Architecture
-  The project has already started migrating to a modular features/ structure, but it can be more rigorous.
-   * Relocate Contexts: Move domain-specific contexts from the root contexts/ folder into their respective feature folders (e.g., features/auth/contexts/AuthContext.tsx and
-     features/habit/contexts/HabitContext.tsx).
-   * Internal Consistency: Ensure every feature folder follows a standard sub-directory structure:
-
-   1     features/[feature-name]/
-   2     ├── components/    # Feature-specific UI
-   3     ├── hooks/         # Feature-specific hooks (e.g., useHabits)
-   4     ├── services/      # Feature-specific API logic
-   5     ├── types/         # Feature-specific interfaces
-   6     ├── contexts/      # Feature-specific state management
-   7     └── utils/         # Feature-specific helper functions
-   * Clean Feature Subfolders: Rename subfolders like features/habit/components/createHabit to features/habit/components/create-habit to maintain naming consistency.
-
-  4. Optimize Root utils/ and services/
-  The root utilities and services should only house truly global, generic logic.
-   * Consolidate Time Helpers: Merge utils/date-time.ts, utils/date-time.ts, and utils/date-time.ts into a single, cohesive utils/date-time.ts file or a utils/date-time/ directory.
-   * Move Domain Logic: Relocate utils/assessmentCore.ts to a feature folder (e.g., features/assessment/utils/) as it likely contains business-specific logic.
-   * Cleanup & Typos: Rename utils/notoficationHelper.ts to notificationHelper.ts and services/storageSerives.ts to storageService.ts.
-
-  5. Shared vs. Feature Components
-  Clearly separating "dumb" UI primitives from "smart" feature components improves reusability.
-   * UI Primitives: Ensure components/ui/ strictly contains generic, stateless components (e.g., StyledButton, StyledInput).
-   * Smart Components: Move components that depend on domain logic (e.g., PricingModal, UpgradeBanner, StatsCard) from components/ui/ into their respective features (e.g., features/billing/components/).
