--- todoTask/today.txt
@@ -1,42 +1,0 @@
-Nimbus Tools Analysis - February 13, 2026
-
-The following tool screens have been identified and updated to the new "Ultra Premium" AppHeader standard:
-
-1. Routine Plan (RoutineScreen.tsx): Templates for life hacks, beauty, wellness, etc.
-2. Done --- Recipe Plan (RecipeScreen.tsx): Curated recipes for nourishment.   
-3. Article Library (ArticleScreen.tsx): Deep dives into healing and neuroscience.
-4. XX --- Scribbles (ScribbleListScreen.tsx): Thought tracking and personal journaling.
-5. XX --- Nourish Plan (MealPlannerScreen.tsx): Daily meal tracking and nutrition summary.
-6. XX --- Future Plan (MealWeeklyViewScreen.tsx): Weekly overview of scheduled meals.
-7. XX --- Meal Creation (MealCreationScreen.tsx): Interface for planning individual or bulk meals.
-8. XX --- Calorie Intake Calculator (CalorieCalculatorScreen.tsx): Estimates daily caloric needs.
-9. XX --- Protein Intake Calculator (ProteinCalculatorScreen.tsx): Specialized protein guidance.
-10. XX --- ody Shape Calculator (BodyShapeCalculatorScreen.tsx): Silhouette analysis tool.
-
-11. Mood Tracker (MoodTracker/): Emotional well-being monitoring.
-12. Product List (ProductListScreen.tsx): Curated product recommendations.
-13. AI Therapy (AItherapyScreen.tsx): AI-driven therapeutic conversations.
-
-All headers now feature the consistent premium typography, back navigation, and contextual right-side actions.
-
-
-
----
-
-Self Care Analysis - February 13, 2026
-
-The following self-care screens have been identified and are candidates for the new "Ultra Premium" AppHeader standard:
-
-1. Meditation (MeditationScreen.tsx): Guided meditations with category filtering and bottom player.
-2. Soundscapes (SoundscapeScreen.tsx): Ambient audio for focus and relaxation.
-3. XX --- Reflections (ReflectionScreen.tsx): Journaling and mood reflection templates.
-4. Workouts (WorkoutScreen.tsx): Exercise library with categories (Cardio, Strength, Yoga, etc.).
-5. Workout Session (WorkoutSessionScreen.tsx): Interactive timer and exercise guidance.
-6. Mental Health Tests (MentalHealthTestScreen.tsx): Directory of psychological assessments.
-7. Mental Health Assessment (MentalHealthGetStartedScreen.tsx): Multi-step questionnaire flow.
-8. Better Sleep (SleepScreen.tsx): Sleep tracking dashboard and aids (currently a full-screen Modal).
-9. Things To Do (ThingsToDoScreen.tsx): Interactive "Self-Love" activity generator (currently a Modal).
-
-
-
-Standardizing these screens will involve replacing custom headers (MeditationHeader, SoundscapeHeader, etc.) with the reusable AppHeader component.
